+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## [Unreleased]
 
+### Added
+
+- feat(redteam): add goal-misalignment plugin for detecting Goodhart's Law vulnerabilities (#6045)
+
 ## [0.119.1] - 2025-10-29
 
 ### Changed